timeseries integrated - up to failing tests
